% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_model.R
\name{compile_model}
\alias{compile_model}
\title{This function translates a model that has been defined in an MCSim model
(".model") file into the C language (i.e., a ".c" file). It then compiles the
model to create an object code (".o") file and a dynamic linked library
(".dll") file, as well as an R script ("_inits.R") containing several R
functions that can be used for initializing model states and parameters.}
\usage{
compile_model(mName)
}
\description{
Inputs:
mName: String containing the name of the MCSim model. Exclude the file name
suffix ".model". If the function is called from a working directory other
than the one containing the ".model" file, the full path of the ".model"
file should be provided.
}
