% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_model.R
\name{load_model}
\alias{load_model}
\title{This function loads a model that is defined using DLL and "_inits.R" files
based on translation of an MCSim model (".model") file. The R functions
"initParms" and "initStates" and the R vector "Outputs" are defined and
assigned to the "global" environment.}
\usage{
load_model(mName)
}
\description{
Inputs:
mName: String containing the name of the MCSim model. Exclude the file name
suffix ".model". If the function is called from a working directory other
than the one containing the ".model" file, the full path of the ".model"
file should be provided.
}
