% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcsim.R
\name{mcsim}
\alias{mcsim}
\title{Conduct MCSim in simulation}
\usage{
mcsim(model, input, dir = ".")
}
\arguments{
\item{model}{a string giving the name of the MCSim model file.}

\item{input}{a string giving the name of the MCSim input file.}

\item{dir}{a character to set the directory of the MCSim model file.}
}
\description{
The function needs both MCSim model and input files to conduct the simulation.
}
