% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{perc_input}
\alias{perc_input}
\title{Example Input}
\usage{
perc_input()
}
\description{
The example input: A four-compartment model of Tetrachloroethylene (PERC)
toxicokinetics. (Bois et al. 1996).
}
\references{
Bois, F. Y., et al. "Population toxicokinetics of tetrachloroethylene."
\emph{Archives of toxicology} 70 (1996): 347-355.
}
