% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{perc_model}
\alias{perc_model}
\title{Example Model}
\usage{
perc_model()
}
\description{
The example model: A four-compartment model of Tetrachloroethylene (PERC)
toxicokinetics. (Bois et al. 1996).
}
\references{
Bois, F. Y., et al. "Population toxicokinetics of tetrachloroethylene."
\emph{Archives of toxicology} 70 (1996): 347-355.
}
